verified that all inputs are required in both frontend and backend

+++ app/api/deletePost/route.js
@@ -5,6 +5,7 @@ import { NextResponse } from "next/server"
 export async function POST(req) {
     try {
         const { id } = await req.json()
+        if (!id || id?.length == 0) throw new Error("something is wrong");
         await connectDb()
         await Post.deleteOne({ id: id })
         return NextResponse.json({ success: true })
